Refactor options flow to a multi-step menu-driven system
- Consolidated options flow logic into `config_flow.py` and removed redundant `options_flow.py`.
- Implemented `async_show_menu` in `MerakiOptionsFlowHandler` for a cleaner UX.
- Categorized settings into 'General', 'Sensors', 'Cameras', and 'Advanced' steps.
- Updated translation strings to follow Home Assistant Sentence Case standards.
- Enhanced dynamic schema generation based on discovered hardware.
- Fixed initialization and coordinator access in the options flow handler.
- Updated `REQUIREMENTS.md` and added unit tests for the new flow.
