+++ docs/REQUIREMENTS.md
@@ -36,7 +36,7 @@
 | The integration should create sensor, device tracker, and switch entities in Home Assistant.                                                                                                            | Included |
 | The integration must properly handle config entries, including setup, unloading, and reloading.                                                                                                        | Included |
 | The integration needs to have a configuration flow to take the API key and org id.                                                                                                                     | Included |
-| The integration must have a native Home Assistant Options Flow to manage settings (e.g., enabling camera entities, device sensors) instead of using a custom React panel, adhering to HA UX paradigms. | Included |
+| The integration must have a native, multi-step Home Assistant Options Flow to manage settings (e.g., enabling camera entities, device sensors) via a menu-driven interface, adhering to HA UX guidelines for complex integrations. | Included |
 | The integration should handle re-authentication of the API key.                                                                                                                                        | Included |
 
 \*\*Mapping Meraki objects to Home Assistant objects
